Keep adaptive execution behind mobpack run seam
