+++ meerkat-cli/Cargo.toml
@@ -52,7 +52,6 @@ meerkat-store = { workspace = true }
 meerkat-runtime = { workspace = true }
 meerkat-tools = { workspace = true }
 meerkat-mob = { workspace = true, optional = true }
-meerkat-mob-adaptive = { workspace = true, optional = true }
 meerkat-mob-pack = { workspace = true, optional = true }
 meerkat-mob-mcp = { workspace = true, optional = true }
 meerkat-rpc = { workspace = true, optional = true, features = ["mob"] }
@@ -100,7 +99,6 @@ memory-store-session = ["meerkat/memory-store-session"]
 # Subsystems
 mob = [
     "dep:meerkat-mob",
-    "dep:meerkat-mob-adaptive",
     "dep:meerkat-mob-pack",
     "dep:meerkat-mob-mcp",
     "dep:meerkat-rpc",
