+++ meerkat-mob-adaptive/BUILD.bazel
@@ -69,9 +69,8 @@ rust_library(
     ],
     srcs = glob(["src/**/*.rs"]),
     visibility = [
-        "//meerkat-cli:__pkg__",
         "//meerkat-mob-adaptive:__pkg__",
-        "//meerkat-mob-mcp:__pkg__",
+        "//meerkat-mob-pack:__pkg__",
     ],
     rustc_env = {
         "CARGO_PKG_VERSION": "0.7.0",
