+++ meerkat-mob-mcp/BUILD.bazel
@@ -49,7 +49,6 @@ rust_library(
         proc_macro = True,
     ),
     deps = [
-        "//meerkat-mob-adaptive:meerkat_mob_adaptive",
         "//meerkat-mob:meerkat_mob",
         "//meerkat:meerkat",
         "//meerkat:meerkat_client_agent_factory_build",
@@ -91,7 +90,6 @@ rust_library(
         proc_macro = True,
     ),
     deps = [
-        "//meerkat-mob-adaptive:meerkat_mob_adaptive_with_runtime_test_support",
         "//meerkat-mob:meerkat_mob_with_runtime_test_support",
         "//meerkat:meerkat_client_agent_factory_build",
         "//meerkat:meerkat_contracts_agent_factory_build",
@@ -143,7 +141,6 @@ rust_test(
         proc_macro_dev = True,
     ),
     deps = [
-        "//meerkat-mob-adaptive:meerkat_mob_adaptive_with_runtime_test_support",
         "//meerkat-mob:meerkat_mob_with_runtime_test_support",
         "//meerkat:meerkat_with_runtime_test_support",
         "//meerkat:meerkat_client_agent_factory_build",
@@ -179,7 +176,6 @@ rust_binary(
         proc_macro = True,
     ),
     deps = [
-        "//meerkat-mob-adaptive:meerkat_mob_adaptive",
         "//meerkat-mob-mcp:meerkat_mob_mcp",
         "//meerkat-mob:meerkat_mob",
         "//meerkat:meerkat",
